+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ## 0.0.10
 
+* Inline line numbers (added support to selectable text also)
+* added copy icon
+* now isSelectableText and lineNumbers both can be true at same time
+* documentation updated
+* example app updated